rename all the classes related to Game to EscaperoomGame
