+++ src/Policies/EscaperoomGamePolicy.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Tipoff\Scheduler\Policies;
 
 use Illuminate\Auth\Access\HandlesAuthorization;
-use Tipoff\Scheduler\Models\Game;
+use Tipoff\Scheduler\Models\EscaperoomGame;
 use Tipoff\Support\Contracts\Models\UserInterface;
 
-class GamePolicy
+class EscaperoomGamePolicy
 {
     use HandlesAuthorization;
 
@@ -17,7 +17,7 @@ class GamePolicy
         return $user->hasPermissionTo('view games') ? true : false;
     }
 
-    public function view(UserInterface $user, Game $game): bool
+    public function view(UserInterface $user, EscaperoomGame $game): bool
     {
         return $user->hasPermissionTo('view games') ? true : false;
     }
@@ -27,22 +27,22 @@ class GamePolicy
         return false;
     }
 
-    public function update(UserInterface $user, Game $game): bool
+    public function update(UserInterface $user, EscaperoomGame $game): bool
     {
         return $user->hasPermissionTo('update games') ? true : false;
     }
 
-    public function delete(UserInterface $user, Game $game): bool
+    public function delete(UserInterface $user, EscaperoomGame $game): bool
     {
         return false;
     }
 
-    public function restore(UserInterface $user, Game $game): bool
+    public function restore(UserInterface $user, EscaperoomGame $game): bool
     {
         return false;
     }
 
-    public function forceDelete(UserInterface $user, Game $game): bool
+    public function forceDelete(UserInterface $user, EscaperoomGame $game): bool
     {
         return false;
     }
