+++ tests/Support/Providers/NovaPackageServiceProvider.php
@@ -6,7 +6,7 @@ namespace Tipoff\Scheduler\Tests\Support\Providers;
 
 use Tipoff\Scheduler\Nova\Block;
 use Tipoff\Scheduler\Nova\EscaperoomSlot;
-use Tipoff\Scheduler\Nova\Game;
+use Tipoff\Scheduler\Nova\EscaperoomGame;
 use Tipoff\Scheduler\Nova\RecurringSchedule;
 use Tipoff\Scheduler\Nova\ScheduleErarser;
 use Tipoff\TestSupport\Providers\BaseNovaPackageServiceProvider;
@@ -15,7 +15,7 @@ class NovaPackageServiceProvider extends BaseNovaPackageServiceProvider
 {
     public static array $packageResources = [
         Block::class,
-        Game::class,
+        EscaperoomGame::class,
         RecurringSchedule::class,
         ScheduleErarser::class,
         EscaperoomSlot::class,
